requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0
pytesseract>=0.3.10
easyocr>=1.7.0
PyMuPDF>=1.23.0
pdf2image>=1.16.3
mysql-connector-python>=8.3.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
scikit-image>=0.22.0
imutils>=0.5.4